src/meningioma_dl/requirements.txt
torch~=1.13.1
numpy~=1.26.4
pandas~=2.2.2
scipy~=1.14.1
plotly~=5.23.0
monai~=1.3.2
python-dotenv~=1.0.1
ray~=2.10.0
fire~=0.6.0
flwr~=1.10.0
scikit-learn~=1.5.1
shortuuid~=1.0.13
nibabel~=5.2.1
matplotlib~=3.9.2
grpcio~=1.66.0
